Refactor: move some stuff around
- Move the Exceptions into their own module
- Move the prompts - which were static functions - into their own module
- Move the shell class into its own module
