+++ cloudbackup/cmd/shell/exceptions.py
@@ -1,0 +1,10 @@
+class CloudBackupApiShellException(Exception):
+    pass
+
+class CloudBackupApiBadParameters(CloudBackupApiShellException):
+    pass
+
+
+class CloudBackupApiBadAuthData(CloudBackupApiShellException):
+    pass
+
